PROJ-2478-React-Native-Setup: Create initial setup for mobile-kit project #2478
This commit adds the initial configuration files for a new mobile project named "mobile-kit". It includes Tailwind CSS setup, Babel configuration, Expo setup with app specifics, and dependencies for nativewind, React Native, and TypeScript. Additionally, it updates the package manager to Yarn and introduces a .gitignore for handling environment-specific and debugging files.

Signed-off-by: Amire Steve <zoe_steve@hotmail.com>